+++ attribute/key_test.go
@@ -109,6 +109,11 @@ func TestEmit(t *testing.T) {
 			want: `["foo","bar"]`,
 		},
 		{
+			name: `test Key.Emit() can emit a string representing self.BYTES`,
+			v:    attribute.BytesValue([]byte("foo")),
+			want: "Zm9v",
+		},
+		{
 			name: `test Key.Emit() can emit a string representing self.EMPTY`,
 			v:    attribute.Value{},
 			want: "",